Renamed DataTable component to ListTable
Added comments to all components
