+++ app/src/Components/Indicator/IndicatorList.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Query } from "react-apollo";
 import IndicatorRepository from './../../repository/IndicatorRepository';
-import DataTable from '../DataTable/DataTable';
+import ListTable from '../ListTable/ListTable';
 import RouterButton from './../../Components/FormInput/RouterButton';
 
 const IndicatorList = (refetch) => (
@@ -18,7 +18,7 @@ const IndicatorList = (refetch) => (
           <div style={{ float: 'right' }}>
             <RouterButton targetLocation='/indicator/new' disabled={false} label="Add new indicator" />
           </div>
-          <DataTable buttons={[{"name": "edit", "function": test}, {"name": "King", "function": king}]} data={data.allIndicators.nodes} />
+          <ListTable buttons={[{"name": "edit", "function": test}, {"name": "King", "function": king}]} data={data.allIndicators.nodes} />
         </div>
       );
     }}
